feat: add ghostty config
Right now I'm on Windows so I'm not actually using ghostty, only
tried it out a bit through WSLg. But I liked it a lot and it'll
probably be the terminal I use if I move to Linux; so I decided
it'd be a good idea to save the config for now.